feature(select-component): scroll factory
